+++ Cargo.toml
@@ -1,47 +1,7 @@
-# THIS FILE IS AUTOMATICALLY GENERATED BY CARGO
-#
-# When uploading crates to the registry Cargo will automatically
-# "normalize" Cargo.toml files for maximal compatibility
-# with all versions of Cargo and also rewrite `path` dependencies
-# to registry (e.g. crates.io) dependencies
-#
-# If you believe there's an error in this file please file an
-# issue against the rust-lang/cargo repository. If you're
-# editing this file be aware that the upstream Cargo.toml
-# will likely look very different (and much more reasonable)
+[workspace]
 
-[package]
-name = "dotenv"
-version = "0.13.0"
-authors = ["Noemi Lapresta <noemi.lapresta@gmail.com>", "Craig Hills <chills@gmail.com>", "Mike Piccolo <mfpiccolo@gmail.com>", "Alice Maz <alice@alicemaz.com>", "Sean Griffin <sean@seantheprogrammer.com>", "Adam Sharp <adam@sharplet.me>"]
-description = "A `dotenv` implementation for Rust"
-homepage = "https://github.com/purpliminal/rust-dotenv"
-readme = "README.md"
-keywords = ["environment", "env", "dotenv", "settings", "config"]
-license = "MIT"
-repository = "https://github.com/purpliminal/rust-dotenv"
-
-[[bin]]
-name = "dotenv"
-required-features = ["cli"]
-[dependencies.clap]
-version = "2"
-optional = true
-
-[dependencies.failure]
-version = "0.1.1"
-features = ["derive"]
-default-features = false
-
-[dependencies.lazy_static]
-version = "1.0.0"
-
-[dependencies.regex]
-version = "1.0"
-[dev-dependencies.tempdir]
-version = "0.3.0"
-
-[features]
-backtrace = ["failure/backtrace"]
-cli = ["clap"]
-default = ["backtrace"]
+members = [
+    "dotenv",
+    "dotenv_codegen",
+    "dotenv_codegen_impl",
+]
