+++ dotenv_codegen/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 
 name = "dotenv_codegen"
-version = "0.11.0"
+version = "0.14.0"
 authors = [
   "Santiago Lapresta <santiago.lapresta@gmail.com>",
   "Craig Hills <chills@gmail.com>",
@@ -16,8 +16,6 @@ homepage = "https://github.com/purpliminal/rust-dotenv"
 repository = "https://github.com/purpliminal/rust-dotenv"
 description = "A `dotenv` implementation for Rust"
 
-[workspace]
-
 [dependencies]
-dotenv_codegen_impl = { version = "0.1", path = "dotenv_codegen_impl" }
-proc-macro-hack = "0.4"
+dotenv_codegen_impl = { version = "0.13", path = "../dotenv_codegen_impl" }
+proc-macro-hack = "0.5"
